Update header and routes to have crossing list and crossing map on separate routes
