+++ frontend/src/Dashboard/CrossingMapPage/CrossingMapPage.scss
@@ -1,0 +1,9 @@
+@import "../../css/flexbox";
+
+.CrossingMapPage {
+  margin: auto 10%;
+}
+
+.CrossingMapPage__map-container {
+  margin-top: 2%;
+}
