+++ frontend/src/FloodsRoutes.js
@@ -4,8 +4,7 @@ import PrivateRoute from './PrivateRoute';
 import Header from './Dashboard/Header/Header';
 import ManageUsers from './Dashboard/ManageUsersPage/ManageUsers';
 import CreateUser from './CreateUser';
-import CrossingUpdates from './Dashboard/CrossingUpdatesPage/CrossingUpdates'
-import CrossingMap from './Map/CrossingMap';
+import CrossingMapPage from './Dashboard/CrossingMapPage/CrossingMapPage'
 import NewStatusUpdate from './NewStatusUpdate';
 import PublicHomepage from './PublicHomepage';
 import auth from './services/gqlAuth';
@@ -24,13 +23,12 @@ class FloodsRoutes extends Component {
       <div>
         <Route path="/" exact component={PublicHomepage} />
         <Route path="/dashboard" render={(props) => <Header currentUser={currentUser} {...props} />} />
-        <Route path="/dashboard/map" component={CrossingMap} currentUser={currentUser}/>
         <PrivateRoute path="/dashboard/users" component={ManageUsers}
           authenticated={auth.isAuthenticated()}
           authorized={auth.roleAuthorized(['floods_community_admin', 'floods_super_admin'])}
           currentUser={currentUser}
         />
-        <PrivateRoute path="/dashboard/crossings" component={CrossingUpdates}
+        <PrivateRoute path="/dashboard/crossings/map" component={CrossingMapPage}
           authenticated={auth.isAuthenticated()}
           authorized={auth.roleAuthorized(['floods_community_editor','floods_community_admin', 'floods_super_admin'])}
           currentUser={currentUser}
